requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
Choco==1.0.5
click==8.2.1
colorama==0.4.6
dnspython==2.8.0
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
